refactor(cmd/influxqld): rename transpilerd to influxqld
This clarifies the transpiler to only function for influxql and to only
implement the 1.x interface for that specific language. This simplifies
the handler and will allow us to prevent non-1.x things from being
included. It also clearly delineates the responsibility of influxqld as
a compatibility layer rather than a new tool.