feat(query/influxql): allow transpilers to expose configuration options
The transpiler interface contains a new `DefaultConfig()` method that
will return an interface to a struct that can be modified through
reflection (or through a type switch if desired). The config can then be
sent back to the Transpile method to customize the transpilation with
options.

The reason why I did it this way instead of having the options on the
Transpiler instance itself was that this prevents the http service from
having to know too much about how to instantiate a Transpiler. The
Transpiler is still a simple object with no configuration so it can be
swapped in and out easily with another implementation. If we created a
new Transpiler object for each request and had the options set there, we
would need specific code for each transpiler implementation that created
the Transpiler itself.

This also adds a `DecodeParams` function that fills in the config struct
using reflection.