feat(query/influxql): support for group by and windowing
The transpiler now supports grouping by tags and basic windowing
support. The window offsets are not supported yet at all.

For windowing, we use the window function to split the points, perform
the aggregate/selector operation, and then we put them back into the
same window so they are within the same table as they originally were
located in. This is now reflected in the spec and the code.